third_party/blink/web_tests/platform/generic/fast/xpath/implicit-node-args-expected.txt

Test for bug 12602:

//div[number() = 123]: SUCCESS
//div[string-length() = 3]: SUCCESS
//div[normalize-space() = '123']: SUCCESS
//div[string() = '123']: SUCCESS
